+++ src/services/temporaryStorage/index.ts
@@ -1,0 +1,2 @@
+export * from './memoryTemporaryStorage.service';
+export * from './temporaryStorage.interface';