+++ src/types/tokens.ts
@@ -3,6 +3,7 @@ import {InjectionToken} from '@angular/core';
 import {StringLocalization, NoStringLocalizationService} from '../services/stringLocalization';
 import {PermanentStorage} from '../services/permanentStorage';
 import {Logger, DummyLoggerService} from '../services/logger';
+import {TemporaryStorage, MemoryTemporaryStorageService} from '../services/temporaryStorage';
 
 /**
  * Base url used during SSR when using HTTP (example: http://localhost:8888/)
@@ -33,3 +34,8 @@ export const STRING_LOCALIZATION: InjectionToken<StringLocalization> = new Injec
  * Token used for injecting permanent storage
  */
 export const PERMANENT_STORAGE: InjectionToken<PermanentStorage> = new InjectionToken<PermanentStorage>('PERMANENT_STORAGE');
+
+/**
+ * Token used for injecting temporary storage
+ */
+export const TEMPORARY_STORAGE: InjectionToken<TemporaryStorage> = new InjectionToken<TemporaryStorage>('TEMPORARY_STORAGE', {providedIn: 'root', factory: () => new MemoryTemporaryStorageService()});